modificado:     lib/CiA402Device 	modificado:     lib/fcontrol 	modificado:     lib/sensor-integration 	modificado:     main/clInc20degstep.cpp 	modificado:     main/encoderfb-fo-yes.cpp 	modificado:     main/imufb-inc-fo-yes.cpp 	modificado:     main/ol-pos-360.cpp 	modificado:     tools/motor-check.cpp 	modificado:     tools/move-motor.cpp 	modificado:     tools/start-pos.cpp 	modificado:     tools/test_sensor.cpp

+++ tools/test_sensor.cpp
@@ -8,7 +8,7 @@
 int main(){
     //--sensor--
     SerialArduino tilt;
-    float incSensor,oriSensor;
+    double incSensor,oriSensor;
     double dts=0.1;
     sleep(4); //wait for sensor
     SystemBlock filterSensor(0.09516,0,- 0.9048,1);
